Update method for role assignment and add retry
